requirements.txt
streamlit
pandas
requests
matplotlib
google-generativeai>=0.7.2